pyproject.toml
[project]
name = "customer support chatbot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "faiss-cpu>=1.11.0",
    "google-generativeai>=0.8.5",
    "numpy>=2.2.6",
    "openai>=1.82.1",
    "pandas>=2.2.3",
    "pdfplumber>=0.11.6",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.45.1",
    "trafilatura>=2.0.0",
    "icecream>=2.1.4",
]
